initial authenticator implemetation
